requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
opencv-python==4.8.0.74
numpy==1.25.2
pillow==10.0.0
pydantic==2.3.0
python-dotenv==1.0.0
google-generativeai==0.3.1